Check android.permission.CAMERA in the package
According to the PR conversation, when android.permission.CAMERA is not set in the package, there is no need to ask for the camera permission. Also, checking now camera and storage permissions separately, so if only one of them is missing, the other one will be requested and not both.