Fix Android bug in permission requests
While trying to pick a picture form the library, the application ask for READ_EXTERNAL_STORAGE at each time.
This bug is due to the fact that if permissions[0] (CAMERA) is not granted, the code asks for READ_EXTERNAL_STORAGE.

Now, READ_EXTERNAL_STORAGE permission (persmissions[1]) is checked and asked if not granted, once.