Double exif rotation on android with camera
This solves the "correctOrientation not rotating / wrong orientation with Android camera bug"

If flag correctOrientation is set the image is rotated according to EXIF-data. After this, the original EXIF-data is added back to the image where the rotation is stil there. Result is that instead of a wrongly rotated image with EXIF-data to correct you end up with a correctly rotated image with EXIF-data that tilts the image 90 degrees.

Removing the EXIF completely solves this for me - there should be a more elegant solution to just remove the "orientation" data from the EXIF to make this a more general purpose solution.