iOS: change variableRect from a preference to a parameter,
so that both cropping methods can be used within one App (e.g. our
variable cropping rectangle for all normal photos, but the standard
square for special portraits for contacts).

With that the icon of the shutter release button of our overlay for
rectangle cropping camera was changed from that camera icon to the
standard round white dot with white circle around; to get compatible to
the square cropping camera.